requirements.txt
botometer==1.6.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-config-file==0.6.0
click-plugins==1.1.1
colorama==0.4.5
configobj==5.0.6
humanize==4.3.0
idna==3.3
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.64.0
twarc==2.11.2
tweepy==3.10.0
urllib3==1.26.12
